Move image classification to own folder
